requirements.txt
PyQt5
pydicom
pynetdicom
Pillow
